+++ modules/matterhorn-index-service/src/main/java/org/opencastproject/index/service/impl/index/event/EventSearchQuery.java
@@ -82,6 +82,7 @@ public class EventSearchQuery extends AbstractSearchQuery {
   private Boolean blacklisted = null;
   private Boolean hasComments = null;
   private Boolean hasOpenComments = null;
+  private Boolean needsCutting = null;
   private final List<String> publications = new ArrayList<String>();
   private String workflowScheduledDate = null;
   private Long archiveVersion = null;
@@ -1047,6 +1048,19 @@ public class EventSearchQuery extends AbstractSearchQuery {
   }
 
   /**
+   * Selects recordings with the given has comment need cutting status.
+   *
+   * @param needsCutting
+   *          the event has the comments status that it needs cutting
+   * @return the enhanced search query
+   */
+  public EventSearchQuery withNeedsCutting(boolean needsCutting) {
+    clearExpectations();
+    this.needsCutting = needsCutting;
+    return this;
+  }
+
+  /**
    * Returns the has comments status of the recording.
    *
    * @return the recording has comments status
@@ -1062,6 +1076,15 @@ public class EventSearchQuery extends AbstractSearchQuery {
    */
   public Boolean getHasOpenComments() {
     return hasOpenComments;
+  }
+
+  /**
+   * Returns the has open comments reason that it needs cutting of the recording.
+   *
+   * @return the event has the open comments status that it needs cutting
+   */
+  public Boolean needsCutting() {
+    return needsCutting;
   }
 
   /**